patch: refactored code
